extract _validate_contract_type_or_function_abi to validation.py
